Refactor autocomplete into a NoteTextEditor view and add it to RepliesView
